requirements.txt
wheel
pandas>=1.2.5
matplotlib>=3.4.2
scipy>=1.7.0
click>=7.1.2
tables>=3.6.1
openpyxl>=3.0.0
cmf==1.6.0
spotpy==1.5.14
sphinx>=4.0.0
networkx>=2.4
pyyaml
